Massive update to the project codebase and structure
The code is more stable and more organized. The project structure and
file names have changed a little bit. TXTRFileType was split into
TXTRFileTypeLib which holds the code for processing TXTRs where as
TXTRFileType itself is now the Paint.NET plugin interface to
TXTRFileTypeLib. libtxtr is introduced, allowing for unmanaged
languages to interop with TXTRFileTypeLib. libtxtrPython introduced,
allowing python code to interface with libtxtr. TXTRFileTypeKrita is
introduced, a plugin interface to TXTRFileTypeLib. Finally,
TXTRFileTypeCLI (txtrtool) is introduced, providing a command line
program to interface with TXTRFileTypeLib.